requirements.txt
PyGithub==1.39
SQLAlchemy==1.2.7
